Use int as the coordinate.

+++ watch/analog-watch.c
@@ -43,7 +43,7 @@ typedef void (*draw_function)(tContext *pContext);
 static void drawFace0(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t sx, sy, ex, ey;
+  int sx, sy, ex, ey;
 
   for(int angle = 0; angle < 359; angle +=30)
   {
@@ -61,7 +61,7 @@ static void drawFace0(tContext *pContext)
 static void drawFace3(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t sx, sy, ex, ey;
+  int sx, sy, ex, ey;
 
   for(int angle = 0; angle < 359; angle += 6)
   {
@@ -86,7 +86,7 @@ static void drawFace3(tContext *pContext)
 static void drawFace6(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t x, y;
+  int x, y;
   GrCircleDraw(pContext, CENTER_X, CENTER_Y, 65);
   GrCircleDraw(pContext, CENTER_X, CENTER_Y, 66);
 
@@ -112,7 +112,7 @@ static void drawFace6(tContext *pContext)
 static void drawFace7(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t x, y;
+  int x, y;
   GrCircleDraw(pContext, CENTER_X, CENTER_Y, 62);
   GrCircleDraw(pContext, CENTER_X, CENTER_Y, 63);
 
@@ -129,7 +129,7 @@ static void drawFace7(tContext *pContext)
 static void drawFace4(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t sx, sy, ex, ey;
+  int sx, sy, ex, ey;
 
   for(int angle = 0; angle < 359; angle += 6)
   {
@@ -154,7 +154,7 @@ static void drawFace4(tContext *pContext)
 static void drawFace1(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t x, y;
+  int x, y;
 
   for(int angle = 0; angle < 359; angle += 6)
   {
@@ -176,7 +176,7 @@ static void drawFace1(tContext *pContext)
 static void drawFace5(tContext *pContext)
 {
   int cos_val, sin_val;
-  uint8_t x, y;
+  int x, y;
 
   GrContextFontSet(pContext, &g_sFontNimbus30);
 
@@ -207,7 +207,7 @@ static void drawHand0(tContext *pContext)
 {
   int cos_val, sin_val;
   int angle;
-  uint16_t x, y;
+  int x, y;
 
   // minute hand = length = 70
   angle = _minute * 6+ _sec /10;
@@ -230,7 +230,7 @@ static void drawHand1(tContext *pContext)
 {
   int cos_val, sin_val;
   int angle;
-  uint16_t sx, sy, ex, ey;
+  int sx, sy, ex, ey;
 
   // draw the circle
   GrCircleFill(pContext, CENTER_X, CENTER_Y, 3);
@@ -269,7 +269,7 @@ static void drawHand4(tContext *pContext)
 {
   int cos_val, sin_val;
   int angle;
-  uint16_t sx, sy, ex, ey, mx, my;
+  int sx, sy, ex, ey, mx, my;
 
   // degree is caculated by 
   // arctan(35/7)=?deg
@@ -332,7 +332,7 @@ static void drawHand5(tContext *pContext)
 {
   int cos_val, sin_val;
   int angle;
-  uint16_t sx, sy, ex, ey, mx, my;
+  int sx, sy, ex, ey, mx, my;
 
   // degree is caculated by 
   // arctan(35/7)=?deg
